+++ _src/interfaces/symbols.ts
@@ -1,0 +1,3 @@
+export class NaniumSymbols {
+	static bufferInternalValueSymbol: symbol = Symbol.for('__Nanium__BufferInternalValueSymbol__');
+}