+++ _src/managers/consumers/browserHttp.ts
@@ -7,6 +7,7 @@ import { EventHandler } from '../../interfaces/eventHandler';
 import { HttpCore } from './http.core';
 import { EventSubscription } from '../../interfaces/eventSubscription';
 import { genericTypesSymbol, NaniumObject, responseTypeSymbol } from '../../objects';
+import { NaniumBuffer } from '../../interfaces/naniumBuffer';
 
 export interface NaniumConsumerBrowserHttpConfig extends ServiceConsumerConfig {
 	apiUrl?: string;
@@ -123,6 +124,8 @@ export class NaniumConsumerBrowserHttp implements ServiceManager {
 										try {
 											if (request.constructor[responseTypeSymbol] === ArrayBuffer) {
 												observer.next(value);
+											} else if (request.constructor[responseTypeSymbol]?.name === NaniumBuffer.name) {
+												observer.next(new NaniumBuffer(value));
 											} else {
 												deserialized = this.config.serializer.deserializePartial(value, restFromLastTime);
 												if (deserialized.data?.length) {
@@ -180,7 +183,7 @@ export class NaniumConsumerBrowserHttp implements ServiceManager {
 		throw new Error('not implemented');
 	}
 
-	async httpRequest(method: 'GET' | 'POST', url: string, body?: string | ArrayBuffer, headers?: any): Promise<ArrayBuffer> {
+	async httpRequest(method: 'GET' | 'POST', url: string, body?: string | ArrayBuffer | FormData, headers?: any): Promise<ArrayBuffer> {
 		return new Promise<ArrayBuffer>((resolve: Function, reject: Function) => {
 			// transmission
 			const abortController: AbortController = new AbortController();
@@ -206,56 +209,6 @@ export class NaniumConsumerBrowserHttp implements ServiceManager {
 				.finally(() => {
 					this.activeRequests = this.activeRequests.filter(r => r !== abortController);
 				});
-		});
-	}
-
-	async httpRequest_old(method: 'GET' | 'POST', url: string, body?: string | ArrayBuffer, headers?: any): Promise<string> {
-		return new Promise<string>((resolve: Function, reject: Function) => {
-			let xhr: XMLHttpRequest;
-			try {
-				xhr = new XMLHttpRequest();
-				this.activeRequests.push(xhr);
-				xhr.onabort = (e) => {
-					this.activeRequests = this.activeRequests.filter(r => r !== xhr);
-					reject(e);
-				};
-				xhr.onerror = (e) => {
-					this.activeRequests = this.activeRequests.filter(r => r !== xhr);
-					reject(e);
-				};
-				xhr.onload = async (): Promise<void> => {
-					if (xhr.status === 200) {
-						this.activeRequests = this.activeRequests.filter(r => r !== xhr);
-						if (xhr.response !== undefined && xhr.response !== '') {
-							resolve(xhr.response);
-						} else {
-							resolve();
-						}
-					} else {
-						this.activeRequests = this.activeRequests.filter(r => r !== xhr);
-						if (xhr.response !== undefined && xhr.response !== '') {
-							reject(xhr.response);
-						} else {
-							reject();
-						}
-					}
-				};
-				xhr.open(method, url);
-				for (const key in headers ?? {}) {
-					if (headers.hasOwnProperty(key)) {
-						xhr.setRequestHeader(key, headers[key]);
-					}
-				}
-				xhr.setRequestHeader('Content-Type', this.config.serializer.mimeType);
-				if (method === 'GET') {
-					xhr.send();
-				} else {
-					xhr.send(body);
-				}
-			} catch (e) {
-				this.activeRequests = this.activeRequests.filter(r => r !== xhr);
-				reject(e);
-			}
 		});
 	}
 }